+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- Add addError and removeError public methods
 - Use webpack to build package
 - Fix imports in usage and example